refactor use Identical node
