+++ rules/type-declaration/tests/Rector/Identical/FlipTypeControlToUseExclusiveTypeRector/FlipTypeControlToUseExclusiveTypeRectorTest.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Rector\TypeDeclaration\Tests\Rector\Assign\FlipTypeControlToUseExclusiveTypeRector;
+namespace Rector\TypeDeclaration\Tests\Rector\Identical\FlipTypeControlToUseExclusiveTypeRector;
 
 use Iterator;
 use Rector\Core\ValueObject\PhpVersionFeature;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
-use Rector\TypeDeclaration\Rector\Assign\FlipTypeControlToUseExclusiveTypeRector;
+use Rector\TypeDeclaration\Rector\Identical\FlipTypeControlToUseExclusiveTypeRector;
 use Symplify\SmartFileSystem\SmartFileInfo;
 
 final class FlipTypeControlToUseExclusiveTypeRectorTest extends AbstractRectorTestCase
